Move general parallel detail place code out of parallel_refine
